requirements.txt
# Core dbt + linters
dbt-core==1.8.1
dbt-snowflake==1.8.1
sqlfluff==3.0.6

# Iceberg support
pyiceberg[s3fs]>=0.5.0

# General utilities
requests>=2.28.0,<3.0.0
python-binance>=1.0.16,<1.1.0
pandas
boto3
python-dotenv

# Airflow providers
apache-airflow-providers-snowflake
apache-airflow-providers-trino==5.7.0
apache-airflow-providers-docker

# Astro tools
astronomer-cosmos
astro-sdk-python
